Move the PSResourceInfo class to its own source file, and rename test
