+++ test/PSResourceInfo.Tests.ps1
